# Tweak type inference to improve TTFX