@oxinabox 2023-06-16
Getting rid of the type parameters makese sense to me.
Compiling optimized code when data download is going to utterly dwarf anything else is too much.

Our tests have kind of bit-rotted.